requirement.txt
asgiref==3.6.0
Django==3.2.17
djangorestframework==3.14.0
IP2Location==8.9.0
Pillow==9.4.0
python-decouple==3.7
pytz==2022.7.1
sqlparse==0.4.3
typing_extensions==4.4.0
